# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bureaucat - Approval workflow application API",
        "title": "Bureaucat API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1341",
    "basePath": "/api/v1",
    "paths": {
        "/admin/import/plane": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload and import a Plane.so SQL dump file. Max 500MB. Requires admin role.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Import"
                ],
                "summary": "Import from Plane.so",
                "parameters": [
                    {
                        "type": "file",
                        "description": "SQL dump file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/sso/{provider}": {
            "get": {
                "description": "Initiates the OAuth2/OIDC flow. Redirects to the identity provider.",
                "tags": [
                    "SSO"
                ],
                "summary": "Start SSO",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SSO provider (google or zitadel)",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sso/{provider}/callback": {
            "get": {
                "description": "Handles the OAuth2/OIDC callback. Redirects to frontend with auth status.",
                "tags": [
                    "SSO"
                ],
                "summary": "SSO callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SSO provider (google or zitadel)",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "OAuth state",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}